Use canvas blitting in matplotlib to significantly reduce the stress of plotting line segments quickly on the gui.
Because that decreases the amount of time a sweep takes to complete to
less than a second for 100MHz span, it's hard to read the live scan
if it clears the line at the end of each sweep, so now clear and replot each bin
separately instead of the whole line. More consistency with an actual SpecAn.

Remove the concept of a live and static graph. Instead, double clicking
the live graph simply pauses graphing. With this simpler model, we don't
need to use a notebook gui layout anymore.