Update MSODTransferSpec.md
Ranga has suggested that we remove all spaces and colons from strings in the transfer spec. Transfer spec v1.0.12 has those changes. I also changed “FFT power” to “DFT”, which is a better descriptor for measurement type in my opinion. We need to start thinking about version control, e.g., if it’s version 1.0.11 then spaces and colons need to be stripped and we need to look for mType “FFT power” instead of “DFT”. I will not make these changes on the VA Beach sensor until we make another significant change to the firmware.